fix: Address second round of review feedback, consolidate shared struct
- Fix inconsistent cleanup in filemenu_close_guestdb: move odbCloseFile
  before filewindowtable removal so early return on failure leaves no
  inconsistent state (claude[bot] critical #2)
- Move tyodblistrecord to odbinternal.h, removing duplicates from
  dbverbs.c, headless_filemenu_verbs.c, and headless_db_verbs.c
  (claude[bot] low #5, cursor[bot])
- Add clarifying comment on menuactivelayer->menuactiveitem field
  mapping in mepackmenustructure_v7 (claude[bot] #1 was false positive:
  different structs with different field names for same semantic value)
- Fix Pascal string length bytes in error messages that prevented
  try/else from catching fileMenu.open errors
- Integration test runner: delete stale .root7 and force fresh migration
  from v6 root before each test run
